+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.2.6
 - Add support for index component files
+- Add support for relative components
 - Allow direct access to attributes underlying hash
 - Fix each statement with a key value pair
 - Improve output of binary HTML attributes